+++ app/src/main/kotlin/ac/mdiq/podcini/storage/parser/ID3Reader.kt
@@ -3,7 +3,6 @@ package ac.mdiq.podcini.storage.parser
 import ac.mdiq.podcini.storage.utils.CountingSource
 import ac.mdiq.podcini.utils.Logd
 import ac.mdiq.podcini.utils.Loge
-import ac.mdiq.podcini.utils.Logt
 import com.fleeksoft.charset.decodeToString
 import okio.Buffer
 import okio.BufferedSource